feat(auth-api): implement email OTP verification for signup

+++ backend/src/auth/auth.service.ts
@@ -13,7 +13,7 @@ import { SignUpDto } from './dto/sign-up-auth.dto';
 
 import * as bcrypt from 'bcrypt';
 import { RefreshTokenDto } from './dto/refresh-token.dto';
-import { User, UserDocument } from 'src/users/schemas/user.schema';
+import { User, UserDocument, UserStatus } from 'src/users/schemas/user.schema';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
 import { RefreshToken } from './schemas/refresh-token.schema';
@@ -21,6 +21,8 @@ import { randomUUID } from 'crypto';
 import { ChangePasswordDto } from './dto/change-password.dto';
 import { ResetToken } from './schemas/reset-token.schema';
 import { MailService } from 'src/services/mail.service';
+import { Otp, OtpDocument } from './schemas/email-otp.schema';
+import { VerifyOtpDto } from './dto/verify-otp.dto';
 
 @Injectable()
 export class AuthService {
@@ -36,6 +38,9 @@ export class AuthService {
     // thêm model User
     @InjectModel(User.name)
     private userModel: Model<UserDocument>,
+
+    @InjectModel(Otp.name)
+    private otpModel: Model<OtpDocument>,
 
     // thêm model RefreshToken
     @InjectModel(RefreshToken.name)
@@ -61,6 +66,12 @@ export class AuthService {
     // nếu không có user hoặc mật khẩu không đúng thì báo lỗi
     if (!user) {
       throw new UnauthorizedException('Email không đúng hoặc không tồn tại');
+    }
+    if (user.status === UserStatus.BLOCKED) {
+      throw new UnauthorizedException('Tài khoản đã bị khóa');
+    }
+    if (!user.isVerified) {
+      throw new UnauthorizedException('Email chưa được xác thực');
     }
 
     /* so sánh password từ client gửi lên với password của hệ thống lưu ở database 
@@ -113,11 +124,21 @@ export class AuthService {
     };
   }
 
+  private generateOtp(): string {
+    return Math.floor(100000 + Math.random() * 900000).toString();
+  }
   //  ============================= đăng ký =============================
   async signUp(signUpDto: SignUpDto) {
     console.log('signUpDto:', signUpDto);
 
-    const { password } = signUpDto;
+    const { email, password } = signUpDto;
+
+    const existedUser = await this.usersService.findByEmail(email);
+
+    if (existedUser) {
+      throw new BadRequestException('Email already exists');
+    }
+
     // =============================bcrypt mật khẩu==================================
     const saltOrRounds = 10;
     // cho password vào hàm bcryp.hash để mã hóa
@@ -125,12 +146,54 @@ export class AuthService {
     const hashedPassword = await bcrypt.hash(password, saltOrRounds);
     // console.log(`Hash: ${hashedPassword}`);
 
-    const user = await this.usersService.create({
+    const user = await this.userModel.create({
       ...signUpDto,
       password: hashedPassword,
+      isVerified: false,
     });
 
-    return { message: 'Đăng ký thành công', user };
+    // tạo mã OTP
+    const otp = this.generateOtp();
+
+    // tạo thời gian hết hạn của otp
+    const expiresAt = new Date();
+    expiresAt.setMinutes(expiresAt.getMinutes() + 5);
+
+    // tạo để lưu vào db
+    await this.otpModel.create({
+      email,
+      otp,
+      expiresAt,
+    });
+
+    // gửi mã otp về mail
+    await this.mailService.sendOtpEmail(email, otp);
+
+    return { message: 'OTP sent to email', user };
+  }
+
+  // ============================= verify OTP  =============================
+  async verifyOtp(verifyOtpDto: VerifyOtpDto) {
+    // nhận email và otp từ client
+    const { email, otp } = verifyOtpDto;
+
+    // tìm và xóa toàn bộ OTP của email đó
+    const record = await this.otpModel.findOneAndDelete({ email, otp });
+
+    if (!record) {
+      throw new BadRequestException('OTP không đúng');
+    }
+
+    if (record.expiresAt < new Date()) {
+      throw new BadRequestException('OTP đã hết hạn');
+    }
+
+    await this.userModel.updateOne({ email }, { isVerified: true });
+
+    // await this.otpModel.deleteMany({ email });
+    return {
+      message: 'Xác thực email thành công',
+    };
   }
 
   // ============================= refresh token  =============================
@@ -210,14 +273,8 @@ export class AuthService {
 
   // ============================= quên mật khẩu =============================
   async forgotPassword(email: string) {
-    // nhận email từ client gửi lên
-
     // check email này có tồn tại hay chưa
     const userExisting = await this.usersService.findByEmail(email);
-
-    if (!userExisting) {
-      throw new NotFoundException();
-    }
 
     // nếu email tồn tại tạo token, thời gian hết hạn token
     if (userExisting) {
@@ -236,13 +293,19 @@ export class AuthService {
 
       // gủi cho email này một link chứa token, gọi từ dịch vụ mail
       await this.mailService.sendPasswordResetEmail(email, resetToken);
+
+      return {
+        message: 'If email exists, reset link sent',
+      };
     }
-    // console.log(email);
-    // return { email };
   }
 
   // ============================= reset mật khẩu =============================
   async resetPassword(resetToken: string, newPassword: string) {
+    if (newPassword.length < 6) {
+      throw new BadRequestException('Password too short');
+    }
+
     //  tìm giá trị reset token document trong database
     const token = await this.resetTokenModel.findOneAndDelete({
       resetPasswordToken: resetToken,
@@ -250,7 +313,9 @@ export class AuthService {
       // expiresAt phải lớn hơn hoặc bằng thời điểm hiện tại
       resetPasswordExpires: { $gte: new Date() },
     });
+
     console.log(token);
+
     if (!token) {
       throw new UnauthorizedException('Invalid link');
     }
